+++ apps/flowbite-svelte-texteditor/src/lib/alignment/AlignmentButton.svelte
@@ -40,7 +40,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[AlignmentButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L83)
+[AlignmentButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L84)
 ## Props
 @prop editor
 @prop alignment