+++ apps/flowbite-svelte-texteditor/src/lib/details/DetailsButtonGroup.svelte
@@ -17,7 +17,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[DetailsButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L110)
+[DetailsButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L111)
 ## Props
 @prop editor
 @prop showToolbar = true