+++ apps/flowbite-svelte-texteditor/src/lib/font/FontButton.svelte
@@ -136,7 +136,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[FontButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L128)
+[FontButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L129)
 ## Props
 @prop editor
 @prop format