+++ apps/flowbite-svelte-texteditor/src/lib/font/FontButtonGroup.svelte
@@ -20,7 +20,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[FontButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L133)
+[FontButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L134)
 ## Props
 @prop editor
 @prop showToolbar = true