+++ apps/flowbite-svelte-texteditor/src/lib/groups/GroupCustom.svelte
@@ -18,7 +18,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[GroupEditorBasicProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L323)
+[GroupEditorBasicProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L324)
 ## Props
 @prop editor
 @prop showToolbar = true