+++ apps/flowbite-svelte-texteditor/src/lib/heading/HeadingButton.svelte
@@ -62,7 +62,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[EditorBasicProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L47)
+[EditorBasicProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L48)
 ## Props
 @prop editor
 @prop tooltipText