+++ apps/flowbite-svelte-texteditor/src/lib/invisible-characters/InvisibleButtonGroup.svelte
@@ -19,7 +19,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[InvisibleButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L193)
+[InvisibleButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L194)
 ## Props
 @prop editor
 @prop showToolbar = true