+++ apps/flowbite-svelte-texteditor/src/lib/list/ListButton.svelte
@@ -37,7 +37,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[ListButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L212)
+[ListButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L213)
 ## Props
 @prop editor
 @prop format