+++ apps/flowbite-svelte-texteditor/src/lib/table/TableButton.svelte
@@ -76,7 +76,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[TableButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L247)
+[TableButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L248)
 ## Props
 @prop editor
 @prop action = 'table'