+++ apps/flowbite-svelte-texteditor/src/lib/task/TaskListButton.svelte
@@ -34,7 +34,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[TaskListButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L273)
+[TaskListButtonProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L274)
 ## Props
 @prop editor
 @prop action