+++ apps/flowbite-svelte-texteditor/src/lib/task/TaskListButtonGroup.svelte
@@ -22,7 +22,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[TaskListButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L277)
+[TaskListButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L278)
 ## Props
 @prop editor
 @prop showToolbar = true