+++ apps/flowbite-svelte-texteditor/src/lib/undo-redo/UndoRedo.svelte
@@ -36,7 +36,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[UndoRedoProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L285)
+[UndoRedoProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L286)
 ## Props
 @prop editor
 @prop action