+++ apps/flowbite-svelte-texteditor/src/lib/undo-redo/UndoRedoButtonGroup.svelte
@@ -17,7 +17,7 @@
 @component
 [Go to docs](https://flowbite-svelte.com/docs/plugins/wysiwyg)
 ## Type
-[UndoRedoButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L292)
+[UndoRedoButtonGroupProps](https://github.com/shinokada/flowbite-svelte-plugins/blob/main/src/lib/types.ts#L293)
 ## Props
 @prop editor
 @prop showToolbar = true